requirements.txt
pip
discord.py
python-dotenv